Try fix compilation error on other platforms

+++ source/blender/editors/undo/ed_undo.cc
@@ -837,7 +837,7 @@ static wmOperatorStatus undo_clear_history_exec(bContext *C, wmOperator *op)
 
   const size_t data_size_after = ED_undosys_total_memory_calc(undo_stack);
   const int64_t data_size_freed = math::max(int64_t(data_size_before) - int64_t(data_size_after),
-                                            0L);
+                                            int64_t(0));
 
   char data_size_str[BLI_STR_FORMAT_INT64_BYTE_UNIT_SIZE];
   BLI_str_format_byte_unit(data_size_str, data_size_freed, true);
